+++ src/features/acf/frontend/common/RenderDate/index.tsx
@@ -2,7 +2,7 @@ import { parseDate, formatDate } from "@/features/common/shared/time";
 import type { JSX } from "react";
 import type { GridRenderCellParams, GridValidRowModel } from "@mui/x-data-grid";
 
-export const RenderDate = <TModel extends GridValidRowModel>({
+export const DataGridRenderDate = <TModel extends GridValidRowModel>({
     value,
 }: GridRenderCellParams<TModel, string | null>): JSX.Element => {
     if (!value) return <span>-</span>;
@@ -11,7 +11,7 @@ export const RenderDate = <TModel extends GridValidRowModel>({
     return <span>{formatDate(date)}</span>;
 };
 
-export const GenericRenderDate = ({
+export const RenderDate = ({
     value,
 }: {
     value: string | null;
